fix: query parameters which are agnostic are not coerced
The JSON-API spec says that the `filter` and `page` query
parameters are agnostic. Coercing the square brackets into
an object is forcing a query type into this module, and
we should remain agnostic where the spec does so.

BREAKING CHANGE: The `filter` query parameter is no longer
mapped to an object if it has square braces.

fix #1